refactor: restructure project layout
- move postgres directory to repository
- add http directory to handler
- rename pageID and pageSize to skip and limit
