+++ Taskfile.yml
@@ -40,14 +40,14 @@ tasks:
 
   migrate:up:
     desc: "Run database migrations"
-    cmd: migrate -path ./internal/adapter/postgres/migrations -database {{.DSN}} -verbose up {{.CLI_ARGS}}
+    cmd: migrate -path ./internal/adapter/repository/postgres/migrations -database {{.DSN}} -verbose up {{.CLI_ARGS}}
     requires:
       vars:
         - DSN
 
   migrate:down:
     desc: "Rollback database migrations"
-    cmd: migrate -path ./internal/adapter/postgres/migrations -database {{.DSN}} -verbose down {{.CLI_ARGS}}
+    cmd: migrate -path ./internal/adapter/repository/postgres/migrations -database {{.DSN}} -verbose down {{.CLI_ARGS}}
     requires:
       vars:
         - DSN
