+++ cmd/http/main.go
@@ -4,9 +4,8 @@ import (
 	"log/slog"
 	"os"
 
-	"github.com/bagashiz/go-pos/internal/adapter/handler"
-	"github.com/bagashiz/go-pos/internal/adapter/postgres"
-	"github.com/bagashiz/go-pos/internal/adapter/postgres/repository"
+	handler "github.com/bagashiz/go-pos/internal/adapter/handler/http"
+	repo "github.com/bagashiz/go-pos/internal/adapter/repository/postgres"
 	"github.com/bagashiz/go-pos/internal/core/service"
 	"github.com/joho/godotenv"
 )
@@ -46,7 +45,7 @@ func main() {
 
 	// Init database
 	slog.Info("Connecting to the database...")
-	db, err := postgres.NewDB()
+	db, err := repo.NewDB()
 	if err != nil {
 		slog.Error("Error initializing database connection", "error", err)
 		os.Exit(1)
@@ -63,7 +62,7 @@ func main() {
 
 	// Dependency injection
 	// User
-	userRepo := repository.NewUserRepository(db)
+	userRepo := repo.NewUserRepository(db)
 	userService := service.NewUserService(userRepo)
 	userHandler := handler.NewUserHandler(userService)
