+++ internal/core/port/user.go
@@ -12,7 +12,7 @@ type UserRepository interface {
 	CreateUser(ctx context.Context, user *domain.User) (*domain.User, error)
 	GetUserByID(ctx context.Context, id uint64) (*domain.User, error)
 	GetUserByEmail(ctx context.Context, email string) (*domain.User, error)
-	ListUsers(ctx context.Context, pageId, pageSize uint64) ([]*domain.User, error)
+	ListUsers(ctx context.Context, skip, limit uint64) ([]*domain.User, error)
 	UpdateUser(ctx context.Context, user *domain.User) (*domain.User, error)
 	DeleteUser(ctx context.Context, id uint64) error
 }
@@ -22,7 +22,7 @@ type UserService interface {
 	Register(ctx context.Context, user *domain.User) (*domain.User, error)
 	Login(ctx context.Context, email, password string) (*domain.User, error)
 	GetUser(ctx context.Context, id uint64) (*domain.User, error)
-	ListUsers(ctx context.Context, pageId, pageSize uint64) ([]*domain.User, error)
+	ListUsers(ctx context.Context, skip, limit uint64) ([]*domain.User, error)
 	UpdateUser(ctx context.Context, user *domain.User) (*domain.User, error)
 	DeleteUser(ctx context.Context, id uint64) error
 }
